Entity Metrics Graphs Report: - Fixed #95 - Fixed the time rendering bug on the graphs where the last time comparison should have been <= (not <)
Snapshots Report:
- Added parsing tables addressed by the query (SELECT, INSERT, UPDATE, etc) from the SQL
- Added multiselect option to the execution range
- Added query type and table type parsing for Cassandra calls

DB Entities Report:
- Added parsing tables addressed by the query (SELECT, INSERT, UPDATE, etc) from the SQL for Queries
- Added multiselect option to the execution range

Flame Graphs Report:
- Added a bunch of new frameworks to color in pretty pretty colors
- Added copy ability for the flame graph element
- Fixed exception that happens with lookups of unknown frameworks that are obfuscated
- Fixed ANTLR mapping coloring because I clearly can't spell
- Changed colors of NewtonSoft JSON mapping

Flowmaps Report:
- Removed extraction of per-minute report as nobody was using it

Configuration Report:
- Added Entity Identifier column to configuration differences to be unambigious about which entity is being compared
- Added parsing for Excluded tiers for BT Scopes

Health Check:
- Added a number of new rules for the first release

+++ ReportObjects/HealthCheck/HealthCheckRuleResult.cs
@@ -16,7 +16,7 @@ namespace AppDynamics.Dexter.ReportObjects
         public string Category { get; set; }
         public string Code { get; set; }
         public string Name { get; set; }
-        public double Grade { get; set; }
+        public int Grade { get; set; }
         public string Description { get; set; }
 
         public DateTime EvaluationTime { get; set; }
